requirements.txt
asgiref==3.8.1
Django==5.0.3
django-filter==24.3
djangorestframework==3.15.2
Faker==33.1.0
Markdown==3.7
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.2
typing_extensions==4.12.2
validate-docbr==1.10.0
